refactor: comment 관련 repository 구조 개선
