+++ src/main/java/io/spring/application/article/ArticleService.java
@@ -14,6 +14,7 @@ import io.spring.core.article.domain.Tag;
 import io.spring.core.article.domain.TagRepository;
 import io.spring.core.user.domain.User;
 import jakarta.validation.Valid;
+import java.util.List;
 import lombok.AllArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
@@ -21,76 +22,74 @@ import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-
 @Service
 @AllArgsConstructor
 public class ArticleService {
-	private final ArticleRepository articleRepository;
-	private final TagRepository tagRepository;
+  private final ArticleRepository articleRepository;
+  private final TagRepository tagRepository;
 
-	@Transactional(readOnly = true)
-	public ArticleResponse getArticle(String slug) {
-		Article article = findBySlug(slug);
+  @Transactional(readOnly = true)
+  public ArticleResponse getArticle(String slug) {
+    Article article = findBySlug(slug);
 
-		return ArticleResponse.from(article);
-	}
+    return ArticleResponse.from(article);
+  }
 
-	@Transactional(readOnly = true)
-	public PaginatedListResponse<ArticleSummaryResponse> getArticles(int page, int size) {
-		Page<Article> articlePage =
-			articleRepository.findAll(PageRequest.of(page, size, Sort.by("createdAt").descending()));
+  @Transactional(readOnly = true)
+  public PaginatedListResponse<ArticleSummaryResponse> getArticles(int page, int size) {
+    Page<Article> articlePage =
+        articleRepository.findAll(PageRequest.of(page, size, Sort.by("createdAt").descending()));
 
-		List<ArticleSummaryResponse> articleSummaryResponses =
-			articlePage.getContent().stream().map(ArticleSummaryResponse::from).toList();
+    List<ArticleSummaryResponse> articleSummaryResponses =
+        articlePage.getContent().stream().map(ArticleSummaryResponse::from).toList();
 
-		PageableResponse<ArticleSummaryResponse> pageableResponse =
-			PageableResponse.of(articlePage.getPageable(), articleSummaryResponses);
+    PageableResponse<ArticleSummaryResponse> pageableResponse =
+        PageableResponse.of(articlePage.getPageable(), articleSummaryResponses);
 
-		return PaginatedListResponse.of(articleSummaryResponses, pageableResponse);
-	}
+    return PaginatedListResponse.of(articleSummaryResponses, pageableResponse);
+  }
 
-	@Transactional
-	public void updateArticle(String slug, User user, UpdateArticleParam updateArticleParam) {
-		Article article = findBySlug(slug);
-		isAuthor(user, article);
-		article.update(
-			updateArticleParam.title(), updateArticleParam.description(), updateArticleParam.body());
-	}
+  @Transactional
+  public void updateArticle(String slug, User user, UpdateArticleParam updateArticleParam) {
+    Article article = findBySlug(slug);
+    isAuthor(user, article);
+    article.update(
+        updateArticleParam.title(), updateArticleParam.description(), updateArticleParam.body());
+  }
 
-	@Transactional
-	public void deleteArticle(String slug, User user) {
-		Article article = findBySlug(slug);
-		isAuthor(user, article);
-		articleRepository.delete(article);
-	}
+  @Transactional
+  public void deleteArticle(String slug, User user) {
+    Article article = findBySlug(slug);
+    isAuthor(user, article);
+    articleRepository.delete(article);
+  }
 
-	Article findBySlug(String slug) {
-		return articleRepository.findBySlug(slug).orElseThrow(ResourceNotFoundException::new);
-	}
+  Article findBySlug(String slug) {
+    return articleRepository.findBySlug(slug).orElseThrow(ResourceNotFoundException::new);
+  }
 
-	private void isAuthor(User user, Article article) {
-		if (!user.getId().equals(article.getAuthor().getId())) {
-			throw new NoAuthorizationException();
-		}
-	}
+  private void isAuthor(User user, Article article) {
+    if (!user.getId().equals(article.getAuthor().getId())) {
+      throw new NoAuthorizationException();
+    }
+  }
 
-	@Transactional
-	public Article createArticle(@Valid NewArticleParam newArticleParam, User creator) {
-		Article article =
-			Article.create(
-				newArticleParam.title(),
-				newArticleParam.description(),
-				newArticleParam.body(),
-				creator);
+  @Transactional
+  public Article createArticle(@Valid NewArticleParam newArticleParam, User creator) {
+    Article article =
+        Article.create(
+            newArticleParam.title(),
+            newArticleParam.description(),
+            newArticleParam.body(),
+            creator);
 
-		articleRepository.save(article);
+    articleRepository.save(article);
 
-		List<Tag> tags =
-			newArticleParam.tagList().stream().map(tag -> Tag.create(tag, article)).toList();
+    List<Tag> tags =
+        newArticleParam.tagList().stream().map(tag -> Tag.create(tag, article)).toList();
 
-		tagRepository.saveAll(tags);
+    tagRepository.saveAll(tags);
 
-		return article;
-	}
+    return article;
+  }
 }
