+++ src/main/java/io/spring/core/article/domain/ArticleRepository.java
@@ -1,16 +1,15 @@
 package io.spring.core.article.domain;
 
+import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 
-import java.util.Optional;
-
 public interface ArticleRepository {
-	Optional<Article> findBySlug(String slug);
+  Optional<Article> findBySlug(String slug);
 
-	Page<Article> findAll(PageRequest createdAt);
+  Page<Article> findAll(PageRequest createdAt);
 
-	void delete(Article article);
+  void delete(Article article);
 
-	Article save(Article article);
+  Article save(Article article);
 }
