+++ src/main/java/io/spring/core/article/infrastructure/JpaArticleRepository.java
@@ -1,9 +1,8 @@
 package io.spring.core.article.infrastructure;
 
 import io.spring.core.article.domain.Article;
-import org.springframework.data.jpa.repository.JpaRepository;
-
 import java.util.Optional;
+import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface JpaArticleRepository extends JpaRepository<Article, String> {
   Optional<Article> findBySlug(String slug);
