+++ src/main/java/io/spring/core/comment/infrastructure/CommentRepositoryImpl.java
@@ -1,0 +1,28 @@
+package io.spring.core.comment.infrastructure;
+
+import io.spring.core.comment.domain.Comment;
+import io.spring.core.comment.domain.CommentRepository;
+import java.util.Optional;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+@RequiredArgsConstructor
+public class CommentRepositoryImpl implements CommentRepository {
+  private final JpaCommentRepository jpaCommentRepository;
+
+  @Override
+  public Comment save(Comment comment) {
+    return jpaCommentRepository.save(comment);
+  }
+
+  @Override
+  public void delete(Comment comment) {
+    jpaCommentRepository.delete(comment);
+  }
+
+  @Override
+  public Optional<Comment> findById(String commentId) {
+    return jpaCommentRepository.findById(commentId);
+  }
+}
