+++ src/main/java/io/spring/core/user/infrastructure/JpaUserRepository.java
@@ -1,10 +1,9 @@
 package io.spring.core.user.infrastructure;
 
 import io.spring.core.user.domain.User;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
 
 @Repository
 public interface JpaUserRepository extends JpaRepository<User, String> {
