+++ src/test/java/io/spring/api/UsersApiTest.java
@@ -12,8 +12,8 @@ import io.spring.api.security.WebSecurityConfig;
 import io.spring.api.user.UsersApi;
 import io.spring.api.user.response.UserPersistResponse;
 import io.spring.application.user.UserService;
-import io.spring.core.user.User;
-import io.spring.core.user.UserRepository;
+import io.spring.core.user.domain.User;
+import io.spring.core.user.domain.UserRepository;
 import io.spring.infrastructure.service.DefaultJwtService;
 import java.util.Map;
 import java.util.Optional;
