+++ src/test/java/io/spring/application/comment/CommentServiceTest.java
@@ -8,11 +8,11 @@ import io.spring.api.comment.request.NewCommentParam;
 import io.spring.api.comment.response.CommentPersistResponse;
 import io.spring.api.exception.NoAuthorizationException;
 import io.spring.api.exception.ResourceNotFoundException;
-import io.spring.core.article.Article;
-import io.spring.core.article.ArticleRepository;
-import io.spring.core.comment.Comment;
-import io.spring.core.comment.CommentRepository;
-import io.spring.core.user.User;
+import io.spring.core.article.domain.Article;
+import io.spring.core.article.domain.ArticleRepository;
+import io.spring.core.comment.domain.Comment;
+import io.spring.core.comment.domain.CommentRepository;
+import io.spring.core.user.domain.User;
 import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
